% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgalgris.R
\docType{package}
\name{cgalgris-package}
\alias{cgalgris-package}
\alias{cgalgris-package,}
\alias{cgalgris}
\title{CGAL for gris}
\description{
CGAL functionality.
}
\details{
\tabular{ll}{
Package: \tab cgalgris\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2015-05-05\cr
License: \tab  \cr
}
}
\references{
\url{https://github.com/mdsumner}, \url{http://www.cgal.org/}
}
\author{
Michael D. Sumner \cr \url{http://staff.acecrc.org.au/~mdsumner} \cr \cr
Maintainer: Michael D. Sumner \email{mdsumner@gmail.com} \cr
}
\keyword{package}
