% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_constraint.R
\name{insert_mesh}
\alias{insert_mesh}
\title{Insert segment constraint}
\usage{
insert_mesh(x, y, v0, v1)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{v0}{segment start index (1-based)}

\item{v1}{segment end index (1-based)}
}
\description{
Insert segment constraint
}
