% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laridae-package.R
\docType{package}
\name{laridae-package}
\alias{laridae}
\alias{laridae-package}
\title{laridae: constrained triangulation and meshing with CGAL}
\description{
A persistent constrained Delaunay triangulation handle backed by CGAL.
See \code{\link[=lari_new]{lari_new()}} for the handle API and \code{\link[=lari_triangulate]{lari_triangulate()}} for the
one-shot form.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/laridae}
  \item Report bugs at \url{https://github.com/hypertidy/laridae/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Mark Padgham [contributor]
}

}
\keyword{internal}
