% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_constraint.R
\name{segment_constraint}
\alias{segment_constraint}
\title{Triangulate segments}
\usage{
segment_constraint(x, ...)
}
\arguments{
\item{x}{anything SC-able}

\item{...}{ignored}
}
\value{
nothing useful, dummy NA
}
\description{
Triangulate segments
}
\examples{
segment_constraint(silicate::minimal_mesh)
}
