% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{segment_constraint_cpp}
\alias{segment_constraint_cpp}
\title{Lines}
\usage{
segment_constraint_cpp(x, y, segment)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{segment}{list of segment pairs (index into x,y)}
}
\description{
Lines
}
