% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tri_xy2}
\alias{tri_xy2}
\title{CGAL vertex index MP version#3}
\usage{
tri_xy2(x, y)
}
\description{
This is the long-hand way, using an iteration over all faces, checking
whether they are finite or not, and accessing the face.vertex info by
dereferencing pointers to each vertex of the face.
}
