% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{insert_constraint}
\alias{insert_constraint}
\title{Insert segment constraint}
\usage{
insert_constraint(x, y, v0, v1)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{segment}{list of segment pairs (index into x,y)}
}
\description{
Insert segment constraint
}
