% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laridae-package.R
\docType{package}
\name{laridae-package}
\alias{laridae}
\alias{laridae-package}
\title{laridae: CGAL Triangulation and Constrained Triangulation}
\description{
Experimental wrapper for CGAL to explore constrained triangulations and
   general edge-decomposition structures for shapes.

CGAL functionality.
}
\references{
\url{http://www.cgal.org/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/laridae}
  \item Report bugs at \url{https://github.com/hypertidy/laridae/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Mark Padgham [contributor]
}



}
\keyword{internal}
\keyword{package}
