% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laridae-package.R
\docType{package}
\name{laridae-package}
\alias{laridae-package}
\alias{laridae-package,}
\alias{laridae}
\title{CGAL for topology in R}
\description{
CGAL functionality.
}
\details{
\tabular{ll}{
Package: \tab laridae\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2015-05-05\cr
License: \tab  \cr
}
}
\references{
\url{https://github.com/mdsumner}, \url{http://www.cgal.org/}
}
\author{
Maintainer: Michael D. Sumner \email{mdsumner@gmail.com} \cr
}
\keyword{package}
