% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{segment_constraint}
\alias{segment_constraint}
\title{Insert segment constraint}
\usage{
segment_constraint(x, y, segment)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{segment}{list of segment pairs (index into x,y)}
}
\description{
Insert segment constraint
}
